repos.conf: shallow clone/sync by default
Virtually all users don't care about the entire repo history. This sets
the default clone and sync to be shallow, i.e., only pulling the most
up-to-date changes. This reduces the load on the repo server and
consumed disk-space/inodes.

Advanced devleopers who require the repo history are unlikely to be
using the supplied `repos.conf`.